qcacmn: Change the data type of message type id
The datatype of wd_msg_type_id was uint16_t which truncates the
upper bytes of msg id.

Changing the datatype of wd_msg_type_id from uint16_t to uint32_t.

+++ wmi/inc/wmi_unified_priv.h
@@ -169,7 +169,7 @@ struct fwdebug {
 #endif /* WLAN_OPEN_SOURCE */
 
 struct wmi_wq_dbg_info {
-	uint16_t wd_msg_type_id;
+	uint32_t wd_msg_type_id;
 	qdf_workqueue_t *wmi_wq;
 	qdf_thread_t *task;
 };
